CIRC-626: Move ClaimItemReturnedRequest construction to controller.
