+++ src/main/java/org/folio/circulation/resources/ClaimItemReturnedResource.java
@@ -1,7 +1,7 @@
 package org.folio.circulation.resources;
 
-import static org.folio.circulation.domain.ClaimItemReturnedRequest.ITEM_CLAIMED_RETURNED_DATE;
 import static org.folio.circulation.support.ItemRepository.noLocationMaterialTypeAndLoanTypeInstance;
+import static org.folio.circulation.support.Result.failed;
 import static org.folio.circulation.support.Result.succeeded;
 import static org.folio.circulation.support.ValidationErrorFailure.singleValidationError;
 
@@ -18,12 +18,17 @@ import org.folio.circulation.support.NoContentResult;
 import org.folio.circulation.support.Result;
 import org.folio.circulation.support.RouteRegistration;
 import org.folio.circulation.support.http.server.WebContext;
+import org.joda.time.DateTime;
 
 import io.vertx.core.http.HttpClient;
+import io.vertx.core.json.JsonObject;
 import io.vertx.ext.web.Router;
 import io.vertx.ext.web.RoutingContext;
 
 public class ClaimItemReturnedResource extends Resource {
+  public static final String ITEM_CLAIMED_RETURNED_DATE = "itemClaimedReturnedDateTime";
+  public static final String COMMENT = "comment";
+
   public ClaimItemReturnedResource(HttpClient client) {
     super(client);
   }
@@ -35,21 +40,21 @@ public class ClaimItemReturnedResource extends Resource {
   }
 
   private void claimItemReturned(RoutingContext routingContext) {
-    validateRequest(routingContext)
-      .after(this::processClaimReturned)
+    createClaimItemReturnedRequest(routingContext)
+      .after(request -> processClaimItemReturned(request, routingContext))
       .thenApply(NoContentResult::from)
       .thenAccept(result -> result.writeTo(routingContext.response()));
   }
 
-  private CompletableFuture<Result<Loan>> processClaimReturned(RoutingContext routingContext) {
+  private CompletableFuture<Result<Loan>> processClaimItemReturned(
+    final ClaimItemReturnedRequest request, RoutingContext routingContext) {
+
     final WebContext context = new WebContext(routingContext);
     final Clients clients = Clients.create(context, client);
 
     final LoanRepository loanRepository = new LoanRepository(clients);
     final ItemRepository itemRepository = noLocationMaterialTypeAndLoanTypeInstance(clients);
     final StoreLoanAndItem storeLoanAndItem = new StoreLoanAndItem(loanRepository, itemRepository);
-
-    final ClaimItemReturnedRequest request = ClaimItemReturnedRequest.from(routingContext);
 
     return succeeded(request)
       .after(req -> loanRepository.getById(req.getLoanId()))
@@ -65,10 +70,19 @@ public class ClaimItemReturnedResource extends Resource {
       .claimItemReturned(request.getComment(), request.getItemClaimedReturnedDateTime()));
   }
 
-  private Result<RoutingContext> validateRequest(RoutingContext routingContext) {
-    return succeeded(routingContext).failWhen(
-      ctx -> succeeded(ctx.getBodyAsJson().getString(ITEM_CLAIMED_RETURNED_DATE) == null),
-      request -> singleValidationError("Item claimed returned date is a required field",
+  private Result<ClaimItemReturnedRequest> createClaimItemReturnedRequest(
+    RoutingContext routingContext) {
+
+    final String loanId = routingContext.pathParam("id");
+    final JsonObject body = routingContext.getBodyAsJson();
+
+    if (!body.containsKey(ITEM_CLAIMED_RETURNED_DATE)) {
+      return failed(singleValidationError("Item claimed returned date is a required field",
         ITEM_CLAIMED_RETURNED_DATE, null));
+    }
+
+    return succeeded(new ClaimItemReturnedRequest(
+      loanId, DateTime.parse(body.getString(ITEM_CLAIMED_RETURNED_DATE)),
+      body.getString(COMMENT)));
   }
 }
